requirements.txt
-e .
aws-cdk.core==1.83.0
aws-cdk.aws_apigatewayv2==1.83.0
aws-cdk.aws-apigatewayv2-integrations==1.83.0
aws-cdk.aws-lambda==1.83.0
path==13.1.0
pytest
